1.0.0.7

+++ src/rpcblockchain.cpp
@@ -57,7 +57,7 @@ double GetDifficulty(const CBlockIndex* blockindex)
 }
 
 
-int64_t GetPoWMHashPS()
+double GetPoWMHashPS()
 {
     if (pindexBest->nHeight >= Params().LastPOWBlock())
     {
