+++ src/rpcserver.h
@@ -104,7 +104,7 @@ extern json_spirit::Value getgenerate(const json_spirit::Array& params, bool fHe
 extern json_spirit::Value setgenerate(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value gethashespersec(const json_spirit::Array& params, bool fHelp);
 
-extern int64_t GetPoWMHashPS();
+extern double GetPoWMHashPS();
 extern double GetPoSKernelPS();
 
 extern std::string HelpRequiringPassphrase();
